requirements.txt
Flask
requests
python-dotenv
Flask-Login
psycopg2-binary
Flask-SQLALchemy